+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Minify Maven Plugin
 
+## 1.7.4
+
+* Update Google Closure Compiler to v20140814 (#71).
+* Add support for Closure Library dependencies sorting (#70).
+* Add option to use default externs provided with the Google Closure Compiler (#67).
+
 ## 1.7.3
 
 * Improve docs: missing information about file order (#45).